requirements.txt
python-twitter==3.5
Django==2.2.2
Jinja2==2.10.1
django-simple-history==2.7.2
mysqlclient==1.4.2.post1
Pygments==2.4.2
tabulate==0.8.3
